Restructured the project and added a progress bar.
