+++ videoparser/__init__.py
@@ -1,0 +1,4 @@
+from . import video, frame
+
+Video = video.Video
+Frame = frame.Frame
